SN-17 AccountController & AccountService

+++ src/main/java/ru/skillbox/socialnetwork/repository/PostRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import ru.skillbox.socialnetwork.model.entity.Post;
 
 @Repository
-public interface PostRepository extends PagingAndSortingRepository<Post, Integer> {
+public interface PostRepository extends PagingAndSortingRepository<Post, Long> {
 
 }
