+++ src/main/resources/db/migration/V1.10__post_comment.sql
@@ -1,9 +1,10 @@
 create table post_comment(
-    id serial NOT NULL PRIMARY KEY,
-    "time" timestamp NOT NULL,
-    post_id int4 NOT NULL REFERENCES post (id),
-    parent_id int4 REFERENCES post_comment (id),
-    author_id int4 NOT NULL REFERENCES person (id),
-    comment_text text,
-    is_blocked int2,
-    is_deleted int2);
+                             id           bigserial NOT NULL PRIMARY KEY,
+                             "time"       timestamp NOT NULL,
+                             post_id      bigint    NOT NULL REFERENCES post (id),
+                             parent_id    bigint REFERENCES post_comment (id),
+                             author_id    bigint    NOT NULL REFERENCES person (id),
+                             comment_text text,
+                             is_blocked   int2,
+                             is_deleted   int2
+);
