+++ src/main/resources/db/migration/V1.11__block_history.sql
@@ -1,7 +1,8 @@
-create table block_history (
-    id serial NOT NULL PRIMARY KEY,
-    "time" timestamp NOT NULL,
-    person_id int4 NOT NULL REFERENCES person (id),
-    post_id int4 NOT NULL REFERENCES post (id),
-    comment_id int4 NOT NULL REFERENCES post_comment (id),
-    action varchar(255));
+create table block_history(
+                              id         bigserial NOT NULL PRIMARY KEY,
+                              "time"     timestamp NOT NULL,
+                              person_id  bigint    NOT NULL REFERENCES person (id),
+                              post_id    bigint    NOT NULL REFERENCES post (id),
+                              comment_id bigint    NOT NULL REFERENCES post_comment (id),
+                              action     varchar(255)
+);
